Refactor: Events v2: Move existing provider code to v1 folder (#10730)
