+++ localstack/services/events/rule.py
@@ -19,11 +19,7 @@ from localstack.aws.api.events import (
     TargetIdList,
     TargetList,
 )
-from localstack.services.events.models_v2 import (
-    Rule,
-    TargetDict,
-    ValidationException,
-)
+from localstack.services.events.models import Rule, TargetDict, ValidationException
 
 TARGET_ID_REGEX = re.compile(r"^[\.\-_A-Za-z0-9]+$")
 TARGET_ARN_REGEX = re.compile(r"arn:[\d\w:\-/]*")
